Updated research.html
Made it so research page displays links to research pages. It is designed so that way each research project has an image, a heading with the projects name, and a "last updated" note so users can see when projects were last updated. Each project is intended to link to a page describing the project along with a list of articles for the achievements of each one.

Right now it uses just the EXcIITED project three times.

If the screen is a small mobile device, it displays a list of projects instead one-by-one.