Removed ExCIITED metadata
removed bot text that displays excitted details on search engines and replaced each page's description with an accurate reflection of the page. However, the twitter links still point to the ExCIITED page because they're not actually hooked up to the UAH website yet, so there's no point linking them to real URLs until they are uploaded.